+++ assets/js/brando/pages.js
@@ -11,6 +11,7 @@ class Pages {
     static onClickExpandButton(e, elem) {
         e.preventDefault();
         $(elem).toggleClass('active');
+        $(elem).find('i').toggleClass('fa-plus').toggleClass('fa-times');
         $('tr.child[data-parent-id=' + $(elem).attr('data-id') + ']').toggleClass('hidden');
     }
 }